fix : moved 'authOptions' into libs folder - 'export authOptions' occured error on vercel deployment
